Quill now uses external tclsh
Previously, Quill assumed the user was using the same tclsh on which
Quill itself was running; this was easy, since Quill was distributed as
a kit.  But it turns out to be more useful if Quill finds the tclsh in
the environment; then, Quill can be built as a star pack, and can
support projects in other versions of Tcl.