Quill updates provided package version
Quill now updates the version number in -quill-require blocks for
packages provided by the project itself.  In quillapp/pkgModules.tcl,
for example, the block contains "package require quill", which is a
local package that is provided for use by others.  By setting the
version number to the project version number automatically, we ensure
that the application sees the version in the project tree, rather than
a stable earlier version in the teapot.