macroset_html(n)
Added the notion of a "macro set", which is a set of macros to be
installed into an instance of macro(n).  Moved all of the manpage(5)
and quilldoc(5) macros that made sense into a new macro set,
macroset_html(n), which is now shared by both modules.

Note that each document format has its own CSS settings; and at some
point in the future these might be customizable by the user.