Implemented 'quill teapot fix'
'quill teapot fix' replaces 'quill teapot create' and 'quill teapot
link'.  Instead of making you run Quill itself with sudo, it outputs a
script that you can run with sudo (or give to someone who can).  This
is safer, and it avoids problems with sudo not making environment
variables available to the command being run.  However, it's not quite
ready for prime time.