Embrace the platform string, part I
Previously, I was using the 'os flavor' as a proxy for the
platform::identify string.  This doesn't work very well.  The 'os
flavor' is about platform conventions; the platform::identify string is
about hardware and OS architectures.

As of now:

* In project.quill, the "app" statement has a -exetype that can be
"kit" or "exe".  This replaces -apptypes.
* When building an app, the executable name always includes the project
version and (for "exe" apps) the platform::identify string.
* When checking dependencies, only the base kits for the current
platform are checked.  For cross-platform, we'll do it differently.

The 'dist' mechanism, and the 'env' and 'config' code related to base
kits still needs to be updated.